merge mohammad commits

+++ app/Http/Controllers/Api/UserController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Api;
 
-use Request;
+use App\Events\User\UserStoreEvent;
 use App\Models\User;
 use App\Http\Resources\DBResource;
 use App\Http\Resources\UserResource;
@@ -10,9 +10,10 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\RoleCollection;
 use App\Http\Resources\UserCollection;
 use App\Repositories\Facades\UserRepo;
-use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Resources\PermissionCollection;
+use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
@@ -23,32 +24,88 @@ class UserController extends Controller
 
     public function store(StoreUserRequest $request)
     {
-        return new UserResource(UserRepo::create($request->all()));
+        $data = $request->except('avatar');
+        $data['password']           =   bcrypt($data['password']);
+        $data['activation_token']   =   str_random(32);
+
+        $createdUser    =   UserRepo::create($data);
+        if ($createdUser === 0) {
+            return new DBResource($createdUser);
+        }
+
+        if ($request->hasFile('avatar')) {
+            $createdUser->addMediaFromRequest('avatar')->toMediaCollection(enum('media.user.avatar'));
+        }
+
+        event(new UserStoreEvent($createdUser));
+        return new UserResource($createdUser);
     }
 
     public function show(User $user)
     {
-        return new UserResource($user);
+        return new UserResource($user->load('avatar'));
     }
 
     public function update(UpdateUserRequest $request, User $user)
     {
-        return new DBResource(UserRepo::update($user, $request->all()));
+        if ($request->hasFile('avatar')) {
+            $user->addMediaFromRequest('avatar')->toMediaCollection(enum('media.user.avatar'));
+        }
+        return new DBResource(UserRepo::update($user, $request->except('avatar')));
     }
 
     public function delete(User $user)
     {
+        if (auth_user()->is($user)) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
         return new DBResource(UserRepo::delete($user));
     }
 
     public function restore(string $user)
     {
-        return new DBResource(UserRepo::delete($user));
+        if (auth_user()->id == $user) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
+        return new DBResource(UserRepo::restore($user));
     }
 
     public function destroy(string $user)
     {
-        return new DBResource(UserRepo::delete($user));
+        if (auth_user()->id == $user) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
+        return new DBResource(UserRepo::destroy($user));
     }
 
     /**
